misc refinements
Realized that I need separate status display for annotator vs agent modes; it didn't seem worth the hassle of adding the status pop-up feature for the annotation status display (why/when would it have huge text?), so I simplified its implementation (no expecting input text as markdown and converting that to html).
Fixed a problem where the focused/active element could be missed because it was hiding inside one or more shadow roots.

Signed-off-by: BareBeaverBat <salisbury.100@buckeyemail.osu.edu>