eliminate unnecessary files from build outputs
e.g. several png's were no longer being used by the README
The remaining png's used by the README did not need to be bulking up the dist/ folder outputs of the build process (which then have to be passed around so that testers/annotators can load the unpacked extension in their browsers)
I've temporarily excluded the EULA pdf because I felt like annotators might object to being asked to click 'accept' for a EULA that had gaps in it (and that Boyuan/Professor-Su weren't even sure was the final form of contract that we'd want).

Signed-off-by: BareBeaverBat <salisbury.100@buckeyemail.osu.edu>